move uuid to new assembly draft
