+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/DCL.Components.Events.asmdef
@@ -1,0 +1,35 @@
+{
+  "name": "DCL.Components.Events",
+  "rootNamespace": "",
+  "references": [
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
+    "GUID:1dd0780aa6be12b428c8005d0bee46b8",
+    "GUID:a881b57670b1d2747a6d7f9e32b63230",
+    "GUID:4973650d2444c4561a15d50f24d91cd9",
+    "GUID:f71b8c619c7cf404ebe3a1fa4c7742fd",
+    "GUID:b1087c5731ff68448a0a9c625bb7e52d",
+    "GUID:85e616515af784fe2928e17f05a0b684",
+    "GUID:c34e38f41494f834abff029ddf82af43",
+    "GUID:cc6bfabbfe87b7949aa248893f89ce37",
+    "GUID:f0dbc731e724975458d07ceec2e4d47a",
+    "GUID:6055be8ebefd69e48b49212b09b47b2f",
+    "GUID:fb529e4714eec4e4da9a3ff3e4da8eee",
+    "GUID:ed14108690fd3eb4cb06e8c9a159b673",
+    "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
+    "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
+    "GUID:1e6b57fe78f7b724e9567f29f6a40c2c",
+    "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+    "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
+    "GUID:4a37c94d8d2dc8242af4c424d0b4a8ae"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": false,
+  "precompiledReferences": [],
+  "autoReferenced": true,
+  "defineConstraints": [],
+  "versionDefines": [],
+  "noEngineReferences": false
+}
