+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/DCL.Components.Events.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 44002c75b9995c34d97b8afc551d32eb
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
