+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/Tests/DCL.Components.Events.Tests.asmdef
@@ -1,5 +1,5 @@
 {
-  "name": "UUIDComponentTests",
+  "name": "DCL.Components.Events.Tests",
   "rootNamespace": "",
   "references": [
     "GUID:e555144732b726c4cba5b0f6337fea75",
@@ -36,6 +36,7 @@
     "GUID:0e39b687cf35c7e40a87f984cdaed9f5",
     "GUID:ee6523961ef177343ad763fcd55c7c46",
     "GUID:c4ef273086976e54ea68f64fac03154d",
+    "GUID:44002c75b9995c34d97b8afc551d32eb",
     "GUID:3b80b0b562b1cbc489513f09fc1b8f69",
     "GUID:ba1fa35f5f69c024fabff38cdd17f569"
   ],
